Move UI classes into separate folder
