+++ libs/importer/src/importers/lastpass/access/ui/ui.ts
@@ -1,5 +1,6 @@
+import { OobResult, OtpResult } from "../models";
+
 import { DuoUi } from "./duo-ui";
-import { OobResult, OtpResult } from "./models";
 
 export abstract class Ui extends DuoUi {
   // To cancel return OtpResult.Cancel, otherwise only valid data is expected.
